dynamic-embeddings/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dynamic-embeddings"
version = "0.1.0"
description = "Dynamic JSON chunking and embedding system with automatic strategy selection"
readme = "README.md"
authors = [
    {name = "Adhir Potdar", email = "adhir.potdar@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"

dependencies = [
    "langchain-core>=0.3.0",
    "langchain-postgres>=0.0.15",
    "langchain-openai>=0.2.0",
    "openai>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg[binary,pool]>=3.2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "numpy>=1.24.0",
    "jsonpath-ng>=1.6.0",
    "tiktoken>=0.7.0",
    "rich>=13.0.0",
    "typer>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/adhir-potdar/dynamic-embeddings"
Repository = "https://github.com/adhir-potdar/dynamic-embeddings.git"
Issues = "https://github.com/adhir-potdar/dynamic-embeddings/issues"

[project.scripts]
dynamic-embeddings = "dynamic_embeddings.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=dynamic_embeddings",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true